// clients/typescript/src/API.ts
export type CreateKudoInput = {
  id?: string | null,
  giverId: string,
  receiverId: string,
  message: string,
  kudoVerb: KudoVerb,
  dataSourceApp: DataSourceApp,
  createdAt?: string | null,
  type?: string | null,
  link: string,
  metadata?: string | null,
};

export enum KudoVerb {
  kudos = "kudos",
  props = "props",
  sparkles = "sparkles",
  fireworks = "fireworks",
}


export enum DataSourceApp {
  github = "github",
  slack = "slack",
  teams = "teams",
  twitter = "twitter",
  other = "other",
}
// ...
export type Kudo = {
  __typename: "Kudo",
  id: string,
  giverId: string,
  receiverId: string,
  message: string,
  kudoVerb: KudoVerb,
  dataSourceApp: DataSourceApp,
  createdAt: string,
  type?: string | null,
  link: string,
  metadata?: string | null,
  giver?: Person | null,
  receiver?: Person | null,
  updatedAt: string,
};
// ...
export type CreatePersonInput = {
  id?: string | null,
  username: string,
  email?: string | null,
  profileUrl?: string | null,
  profileImageUrl?: string | null,
  dataSourceApp: DataSourceApp,
};
// ...
export type CreateKudoMutationVariables = {
  input: CreateKudoInput,
  condition?: ModelKudoConditionInput | null,
};

export type CreateKudoMutation = {
  createKudo?:  {
    __typename: "Kudo",
    id: string,
    giverId: string,
    receiverId: string,
    message: string,
    kudoVerb: KudoVerb,
    dataSourceApp: DataSourceApp,
    createdAt: string,
    type?: string | null,
    link: string,
    metadata?: string | null,
    giver?:  {
      __typename: "Person",
      id: string,
      username: string,
      email?: string | null,
      profileUrl?: string | null,
      profileImageUrl?: string | null,
      dataSourceApp: DataSourceApp,
      kudosGiven?:  {
        __typename: "ModelKudoConnection",
        items:  Array< {
          __typename: "Kudo",
          id: string,
          giverId: string,
          receiverId: string,
          message: string,
          kudoVerb: KudoVerb,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          type?: string | null,
          link: string,
          metadata?: string | null,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      kudosReceived?:  {
        __typename: "ModelKudoConnection",
        items:  Array< {
          __typename: "Kudo",
          id: string,
          giverId: string,
          receiverId: string,
          message: string,
          kudoVerb: KudoVerb,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          type?: string | null,
          link: string,
          metadata?: string | null,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    receiver?:  {
      __typename: "Person",
      id: string,
      username: string,
      email?: string | null,
      profileUrl?: string | null,
      profileImageUrl?: string | null,
      dataSourceApp: DataSourceApp,
      kudosGiven?:  {
        __typename: "ModelKudoConnection",
        items:  Array< {
          __typename: "Kudo",
          id: string,
          giverId: string,
          receiverId: string,
          message: string,
          kudoVerb: KudoVerb,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          type?: string | null,
          link: string,
          metadata?: string | null,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      kudosReceived?:  {
        __typename: "ModelKudoConnection",
        items:  Array< {
          __typename: "Kudo",
          id: string,
          giverId: string,
          receiverId: string,
          message: string,
          kudoVerb: KudoVerb,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          type?: string | null,
          link: string,
          metadata?: string | null,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    updatedAt: string,
  } | null,
};
// ...
export type CreatePersonMutationVariables = {
  input: CreatePersonInput,
  condition?: ModelPersonConditionInput | null,
};

export type CreatePersonMutation = {
  createPerson?:  {
    __typename: "Person",
    id: string,
    username: string,
    email?: string | null,
    profileUrl?: string | null,
    profileImageUrl?: string | null,
    dataSourceApp: DataSourceApp,
    kudosGiven?:  {
      __typename: "ModelKudoConnection",
      items:  Array< {
        __typename: "Kudo",
        id: string,
        giverId: string,
        receiverId: string,
        message: string,
        kudoVerb: KudoVerb,
        dataSourceApp: DataSourceApp,
        createdAt: string,
        type?: string | null,
        link: string,
        metadata?: string | null,
        giver?:  {
          __typename: "Person",
          id: string,
          username: string,
          email?: string | null,
          profileUrl?: string | null,
          profileImageUrl?: string | null,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          updatedAt: string,
        } | null,
        receiver?:  {
          __typename: "Person",
          id: string,
          username: string,
          email?: string | null,
          profileUrl?: string | null,
          profileImageUrl?: string | null,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          updatedAt: string,
        } | null,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
    kudosReceived?:  {
      __typename: "ModelKudoConnection",
      items:  Array< {
        __typename: "Kudo",
        id: string,
        giverId: string,
        receiverId: string,
        message: string,
        kudoVerb: KudoVerb,
        dataSourceApp: DataSourceApp,
        createdAt: string,
        type?: string | null,
        link: string,
        metadata?: string | null,
        giver?:  {
          __typename: "Person",
          id: string,
          username: string,
          email?: string | null,
          profileUrl?: string | null,
          profileImageUrl?: string | null,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          updatedAt: string,
        } | null,
        receiver?:  {
          __typename: "Person",
          id: string,
          username: string,
          email?: string | null,
          profileUrl?: string | null,
          profileImageUrl?: string | null,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          updatedAt: string,
        } | null,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type ListKudosQueryVariables = {
  filter?: ModelKudoFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListKudosQuery = {
  listKudos?:  {
    __typename: "ModelKudoConnection",
    items:  Array< {
      __typename: "Kudo",
      id: string,
      giverId: string,
      receiverId: string,
      message: string,
      kudoVerb: KudoVerb,
      dataSourceApp: DataSourceApp,
      createdAt: string,
      type?: string | null,
      link: string,
      metadata?: string | null,
      giver?:  {
        __typename: "Person",
        id: string,
        username: string,
        email?: string | null,
        profileUrl?: string | null,
        profileImageUrl?: string | null,
        dataSourceApp: DataSourceApp,
        kudosGiven?:  {
          __typename: "ModelKudoConnection",
          nextToken?: string | null,
        } | null,
        kudosReceived?:  {
          __typename: "ModelKudoConnection",
          nextToken?: string | null,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      receiver?:  {
        __typename: "Person",
        id: string,
        username: string,
        email?: string | null,
        profileUrl?: string | null,
        profileImageUrl?: string | null,
        dataSourceApp: DataSourceApp,
        kudosGiven?:  {
          __typename: "ModelKudoConnection",
          nextToken?: string | null,
        } | null,
        kudosReceived?:  {
          __typename: "ModelKudoConnection",
          nextToken?: string | null,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};
// ...
export type ListPeopleQuery = {
  listPeople?:  {
    __typename: "ModelPersonConnection",
    items:  Array< {
      __typename: "Person",
      id: string,
      username: string,
      email?: string | null,
      profileUrl?: string | null,
      profileImageUrl?: string | null,
      dataSourceApp: DataSourceApp,
      kudosGiven?:  {
        __typename: "ModelKudoConnection",
        items:  Array< {
          __typename: "Kudo",
          id: string,
          giverId: string,
          receiverId: string,
          message: string,
          kudoVerb: KudoVerb,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          type?: string | null,
          link: string,
          metadata?: string | null,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      kudosReceived?:  {
        __typename: "ModelKudoConnection",
        items:  Array< {
          __typename: "Kudo",
          id: string,
          giverId: string,
          receiverId: string,
          message: string,
          kudoVerb: KudoVerb,
          dataSourceApp: DataSourceApp,
          createdAt: string,
          type?: string | null,
          link: string,
          metadata?: string | null,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type KudosByDateQueryVariables = {
  type: string,
  createdAt?: ModelStringKeyConditionInput | null,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelKudoFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type KudosByDateQuery = {
  kudosByDate?:  {
    __typename: "ModelKudoConnection",
    items:  Array< {
      __typename: "Kudo",
      id: string,
      giverId: string,
      receiverId: string,
      message: string,
      kudoVerb: KudoVerb,
      dataSourceApp: DataSourceApp,
      createdAt: string,
      type?: string | null,
      link: string,
      metadata?: string | null,
      giver?:  {
        __typename: "Person",
        id: string,
        username: string,
        email?: string | null,
        profileUrl?: string | null,
        profileImageUrl?: string | null,
        dataSourceApp: DataSourceApp,
        kudosGiven?:  {
          __typename: "ModelKudoConnection",
          nextToken?: string | null,
        } | null,
        kudosReceived?:  {
          __typename: "ModelKudoConnection",
          nextToken?: string | null,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      receiver?:  {
        __typename: "Person",
        id: string,
        username: string,
        email?: string | null,
        profileUrl?: string | null,
        profileImageUrl?: string | null,
        dataSourceApp: DataSourceApp,
        kudosGiven?:  {
          __typename: "ModelKudoConnection",
          nextToken?: string | null,
        } | null,
        kudosReceived?:  {
          __typename: "ModelKudoConnection",
          nextToken?: string | null,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};